Backend/requirements.txt
azure-functions
requests
openai
firebase-admin